add and use frontdoor config
